requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pymongo>=4.6.1
motor>=3.3.2
pydantic>=2.6.1
pydantic-core>=2.16.2
pydantic-settings>=2.1.
email-validator>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
python-multipart>=0.0.7
groq>=0.4.0
typing-extensions>=4.10.0
instructor
pypdf2
python-docx
beautifulsoup4
